Tolerate unknown fields in CLI command requests
For stream consumer groups CLI commands. This is useful in case
new fields are needed in further versions. A new version
node can ask for new fields to an old version node. If
the old version node returns a known value for unknown fields
instead of failing, the new node can set up appropriate
default value for these fields in the result of the CLI commands.

References #3753